replaced deprecated interface with moved version, delete internal index file
