+++ packages/live-share-canvas/src/core/InkingManager.ts
@@ -51,7 +51,7 @@ import {
     computeQuadPath,
     renderQuadPathToSVG,
 } from "./internals";
-import { IEvent } from "@fluidframework/common-definitions";
+import { IEvent } from "@fluidframework/core-interfaces";
 
 /**
  * Defines available inking tools.
