+++ packages/live-share/src/TimestampProvider.ts
@@ -3,7 +3,7 @@
  * Licensed under the Microsoft Live Share SDK License.
  */
 
-import { ILiveShareHost, INtpTimeInfo, ITimestampProvider } from "./interfaces";
+import { INtpTimeInfo, ITimestampProvider } from "./interfaces";
 
 const IMPROVE_ACCURACY_INTERVAL = 5 * 1000;
 const IMPROVE_ACCURACY_ATTEMPTS = 5;
