+++ packages/live-share/src/internals/LiveObjectManager.ts
@@ -9,19 +9,18 @@ import { IRuntimeSignaler } from "../LiveEventScope";
 import { LiveShareRuntime } from "../LiveShareRuntime";
 import { IContainerRuntimeSignaler, ILiveEvent } from "../interfaces";
 import { LiveEvent } from "../LiveEvent";
+import { IAzureAudience } from "@fluidframework/azure-client";
+import { isILiveEvent } from "./type-guards";
+import { ObjectSynchronizerEvents } from "./consts";
+import { cloneValue, waitUntilConnected } from "./utils";
+import { ContainerSynchronizer } from "./ContainerSynchronizer";
 import {
-    ContainerSynchronizer,
     GetAndUpdateStateHandlers,
     IContainerLiveObjectStoreEvents,
     ILiveClientEventMap,
     ILiveObjectStore,
-    ObjectSynchronizerEvents,
     StateSyncEventContent,
-    cloneValue,
-    isILiveEvent,
-    waitUntilConnected,
-} from "./";
-import { IAzureAudience } from "@fluidframework/azure-client";
+} from "./internal-interfaces";
 
 /**
  * @hidden
