--- packages/live-share/src/internals/index.ts
@@ -1,20 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the Microsoft Live Share SDK License.
- */
-
-export * from "./consts";
-export * from "./LiveShareTokenProvider";
-export * from "./RequestCache";
-export * from "./RoleVerifier";
-export * from "./utils";
-export * from "./LiveShareHostDecorator";
-export * from "./BackwardsCompatibilityHostDecorator";
-export * from "./Deferred";
-export * from "./type-guards";
-export * from "./internal-interfaces";
-export * from "./LiveObjectManager";
-export * from "./ContainerSynchronizer";
-export * from "./FormatFixHostDecorator";
-
-// TODO: delete internals/index.ts
