+++ packages/live-share/src/internals/internal-interfaces.ts
@@ -3,7 +3,7 @@ import {
     ILiveEvent,
     UpdateSynchronizationState,
 } from "../interfaces";
-import { IEvent } from "@fluidframework/common-definitions";
+import { IEvent } from "@fluidframework/core-interfaces";
 
 /**
  * @hidden
